% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{apiCall,KorAPConnection-method}
\alias{apiCall,KorAPConnection-method}
\alias{apiCall}
\title{Internal API call method}
\usage{
\S4method{apiCall}{KorAPConnection}(
  kco,
  url,
  json = TRUE,
  getHeaders = FALSE,
  cache = kco@cache,
  timeout = kco@timeout
)
}
\arguments{
\item{kco}{KorAPConnection object}

\item{url}{request url}

\item{json}{logical that determines if JSON result is expected}

\item{getHeaders}{logical that determines if headers and content should be returned (as a list)}
}
\description{
Internal API call method
}
\keyword{internal}
