% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{defaultAssociationScoreFunctions}
\alias{defaultAssociationScoreFunctions}
\title{Default association score functions}
\usage{
defaultAssociationScoreFunctions()
}
\value{
list of default association score functions
}
\description{
Default association score functions
}
\examples{
\donttest{
new("KorAPConnection", verbose = TRUE) \%>\%
collocationScoreQuery("Perlen", c("verziertes", "Säue"),
  scoreFunctions = append(associationScoreFunctions(),
     list(localMI = function(O1, O2, O, N, E, window_size) {
                       O * log2(O/E)
                    })))
}

}
\seealso{
Other association-score-functions: 
\code{\link{ll}()},
\code{\link{logDice}()},
\code{\link{mi2}()},
\code{\link{mi3}()},
\code{\link{pmi}()}
}
\concept{association-score-functions}
