% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{format_eta_display}
\alias{format_eta_display}
\title{Format ETA information for display}
\usage{
format_eta_display(eta_seconds, estimated_completion_time)
}
\arguments{
\item{eta_seconds}{numeric ETA in seconds (can be NA)}

\item{estimated_completion_time}{POSIXct estimated completion time (can be NA)}
}
\value{
character string with formatted ETA or empty string if NA
}
\description{
Helper function to format ETA information consistently across different methods.
}
\keyword{internal}
