% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggplotly}
\alias{ggplotly}
\title{Convert ggplot2 to plotly with hyperlinks to KorAP queries}
\usage{
ggplotly(g = ggplot2::last_plot(), tooltip = c("x", "y", "url"), ...)
}
\description{
Converts ...
}
\examples{
library(ggplot2)
kco <- new("KorAPConnection", verbose=TRUE)
g <- expand_grid(condition = c("textDomain = /Wirtschaft.*/", "textDomain != /Wirtschaft.*/"),
            year = (2002:2018)) \%>\%
  cbind(frequencyQuery(kco, "[tt/l=Heuschrecke]",
                       paste0(.$condition," & pubDate in ", .$year)))  \%>\%
  ipm() \%>\%
  ggplot(aes(year, ipm, fill = condition, color = condition)) +
##  theme_light(base_size = 20) +
  geom_freq_by_year_ci()
p <- ggplotly(g)
print(p)
## saveWidget(p, paste0(tmpdir(), "heuschrecke.html")


}
