% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{initialize,KorAPConnection-method}
\alias{initialize,KorAPConnection-method}
\title{Initialize KorAPConnection object}
\usage{
\S4method{initialize}{KorAPConnection}(
  .Object,
  KorAPUrl = defaultKorAPUrl(),
  apiVersion = "v1.0",
  apiUrl,
  accessToken = getAccessToken(KorAPUrl),
  oauthClient = NULL,
  oauthScope = "search match_info",
  authorizationSupported = TRUE,
  userAgent = "R-KorAP-Client",
  timeout = 240,
  verbose = FALSE,
  cache = TRUE
)
}
\description{
Initialize KorAPConnection object
}
\keyword{internal}
