% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPCorpusStats.R
\docType{class}
\name{KorAPCorpusStats-class}
\alias{KorAPCorpusStats-class}
\alias{show,KorAPCorpusStats-method}
\title{Class KorAPCorpusStats}
\usage{
\S4method{show}{KorAPCorpusStats}(object)
}
\arguments{
\item{object}{KorAPCorpusStats object}
}
\description{
\code{KorAPCorpusStats} objects can hold information about a corpus or virtual corpus.
\code{KorAPCorpusStats} objects can be obtained by the \code{\link[=corpusStats]{corpusStats()}} method.
}
\section{Slots}{

\describe{
\item{\code{vc}}{definition of the virtual corpus}

\item{\code{tokens}}{number of tokens}

\item{\code{documents}}{number of documents}

\item{\code{sentences}}{number of sentences}

\item{\code{paragraphs}}{number of paragraphs}
}}

