% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPCorpusStats.R
\name{KorAPCorpusStats}
\alias{KorAPCorpusStats}
\title{\code{KorAPCorpusStats}}
\usage{
KorAPCorpusStats(con, vc = NA, query = NA)
}
\arguments{
\item{con}{object obtained from \code{\link{KorAPConnection}}, that contains all necessary connection information}

\item{vc}{string describing the virtual corpus. An empty string (default) means the whole corpus, as far as it is license-wise accessible.}

\item{query}{qquery object returned from \code{\link{KorAPQuery}}}
}
\description{
\code{KorAPCorpusStats}
}
