% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPQuery.R
\name{KorAPFetchAll}
\alias{KorAPFetchAll}
\title{Fetch all results of a KorAP query.}
\usage{
KorAPFetchAll(queryObject, verbose = FALSE)
}
\arguments{
\item{queryObject}{object obtained from \code{\link{KorAPQuery}}}
}
\value{
The \code{queryObject} input parameter with updated fields \code{$collectedMatches}, \code{$matches} (latest bunch only), \code{$nextStartIndex}, \code{$hasMoreMatches}
}
\description{
Fetch all results of a KorAP query.
}
\examples{
q <- KorAPFetchAll(KorAPQuery(KorAPConnection(), "Ameisenplage"))
q$collectedMatches

}
\references{
\url{https://ids-pub.bsz-bw.de/frontdoor/index/index/docId/9026}
}
\seealso{
\code{\link{KorAPFetchRest}}, \code{\link{KorAPFetchNext}}
}
