% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPQuery.R
\name{KorAPFetchNext}
\alias{KorAPFetchNext}
\title{Fetch the next bunch of results of a KorAP query.}
\usage{
KorAPFetchNext(queryObject, offset = queryObject$nextStartIndex,
  maxFetch = maxResultsPerPage, verbose = FALSE)
}
\arguments{
\item{queryObject}{object obtained from \code{\link{KorAPQuery}}}

\item{offset}{start offset for query results to fetch}

\item{maxFetch}{maximum number of query results to fetch}

\item{verbose}{print progress information if true}
}
\value{
The \code{queryObject} input parameter with updated fields \code{$collectedMatches}, \code{$matches} (latest bunch only), \code{$nextStartIndex}, , \code{$hasMoreMatches}
}
\description{
Fetch the next bunch of results of a KorAP query.
}
\examples{
q <- KorAPFetchNext(KorAPQuery(KorAPConnection(), "Ameisenplage"))

}
\references{
\url{https://ids-pub.bsz-bw.de/frontdoor/index/index/docId/9026}
}
\seealso{
\code{\link{KorAPFetchRest}}, \code{\link{KorAPFetchAll}}
}
