% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPQuery.R
\name{KorAPFetchRest}
\alias{KorAPFetchRest}
\title{Fetches all remaining results of a KorAP query.}
\usage{
KorAPFetchRest(queryObject, verbose = FALSE)
}
\arguments{
\item{queryObject}{object obtained from \code{\link{KorAPQuery}}}
}
\value{
The \code{queryObject} input parameter with updated fields \code{$collectedMatches}, \code{$matches} (latest bunch only), \code{$nextStartIndex}, \code{$hasMoreMatches}
}
\description{
Fetches all remaining results of a KorAP query.
}
\examples{
q <- KorAPFetchRest(KorAPQueryNext(KorAPQuery(KorAPConnection(), "Ameisenplage")))
q$collectedMatches

}
\references{
\url{https://ids-pub.bsz-bw.de/frontdoor/index/index/docId/9026}
}
\seealso{
\code{\link{KorAPFetchAll}}, \code{\link{KorAPFetchNext}}
}
