% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{ll}
\alias{ll}
\title{Log likelihood}
\usage{
ll(O1, O2, O, N, E, window_size)
}
\arguments{
\item{O1}{observed absolute frequency of node}

\item{O2}{observed absolute frequency of collocate}

\item{O}{observed absolute frequency of collocation}

\item{N}{corpus size}

\item{E}{expected absolute frequency of collocation (already adjusted to window size)}

\item{window_size}{total window size around node (left neighbour count + right neighbour count)}
}
\description{
Log likelihood
}
\references{
Dunning, T. (1993): Accurate methods for the statistics of surprise and coincidence. Comput. Linguist. 19, 1 (March 1993), 61-74.

Evert, Stefan (2004): The Statistics of Word Cooccurrences: Word Pairs and Collocations. PhD dissertation, IMS, University of Stuttgart. Published in 2005, URN urn:nbn:de:bsz:93-opus-23714.
Free PDF available from <http://purl.org/stefan.evert/PUB/Evert2004phd.pdf>
}
\seealso{
Other association-score-functions: 
\code{\link{defaultAssociationScoreFunctions}()},
\code{\link{logDice}()},
\code{\link{mi2}()},
\code{\link{mi3}()},
\code{\link{pmi}()}
}
\concept{association-score-functions}
