% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{logDice}
\alias{logDice}
\title{log-Dice coefficient}
\usage{
logDice(O1, O2, O, N, E, window_size)
}
\arguments{
\item{O1}{observed absolute frequency of node}

\item{O2}{observed absolute frequency of collocate}

\item{O}{observed absolute frequency of collocation}

\item{N}{corpus size}

\item{E}{expected absolute frequency of collocation (already adjusted to window size)}

\item{window_size}{total window size around node (left neighbour count + right neighbour count)}
}
\description{
log-Dice coefficient
}
\examples{

}
\references{
Rychlý, Pavel (2008): <a href="http://www.fi.muni.cz/usr/sojka/download/raslan2008/13.pdf">A lexicographer-friendly association score.</a> In Proceedings of Recent Advances in Slavonic Natural Language Processing, RASLAN, 6–9.
}
\seealso{
Other association-score-functions: 
\code{\link{defaultAssociationScoreFunctions}()},
\code{\link{ll}()},
\code{\link{mi2}()},
\code{\link{mi3}()},
\code{\link{pmi}()}
}
\concept{association-score-functions}
