% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{mi2}
\alias{mi2}
\title{Pointwise mutual information squared}
\usage{
mi2(O1, O2, O, N, E, window_size)
}
\arguments{
\item{O1}{observed absolute frequency of node}

\item{O2}{observed absolute frequency of collocate}

\item{O}{observed absolute frequency of collocation}

\item{N}{corpus size}

\item{E}{expected absolute frequency of collocation (already adjusted to window size)}

\item{window_size}{total window size around node (left neighbour count + right neighbour count)}
}
\description{
Pointwise mutual information squared
}
\details{
Also referenced to as mutual dependency (MD)
}
\seealso{
Other association-score-functions: 
\code{\link{defaultAssociationScoreFunctions}()},
\code{\link{ll}()},
\code{\link{logDice}()},
\code{\link{mi3}()},
\code{\link{pmi}()}
}
\concept{association-score-functions}
