% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{mi3}
\alias{mi3}
\title{Pointwise mutual information cubed}
\usage{
mi3(O1, O2, O, N, E, window_size)
}
\arguments{
\item{O1}{observed absolute frequency of node}

\item{O2}{observed absolute frequency of collocate}

\item{O}{observed absolute frequency of collocation}

\item{N}{corpus size}

\item{E}{expected absolute frequency of collocation (already adjusted to window size)}

\item{window_size}{total window size around node (left neighbour count + right neighbour count)}
}
\description{
Pointwise mutual information cubed
}
\details{
Also referenced to as log-frequency biased mutual dependency (LFMD)
}
\references{
Daille, B. (1994): Approche mixte pour l’extraction automatique de terminologie: statistiques lexicales et filtres linguistiques. PhD thesis, Université Paris 7.

Thanopoulos, A., Fakotakis, N., Kokkinakis, G. (2002): Comparative evaluation of collocation extraction metrics. In: Proc. of LREC 2002: 620–625.
}
\seealso{
Other association-score-functions: 
\code{\link{defaultAssociationScoreFunctions}()},
\code{\link{ll}()},
\code{\link{logDice}()},
\code{\link{mi2}()},
\code{\link{pmi}()}
}
\concept{association-score-functions}
