% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-scores.R
\name{pmi}
\alias{pmi}
\title{Pointwise mutual information}
\usage{
pmi(O1, O2, O, N, E, window_size)
}
\arguments{
\item{O1}{observed absolute frequency of node}

\item{O2}{observed absolute frequency of collocate}

\item{O}{observed absolute frequency of collocation}

\item{N}{corpus size}

\item{E}{expected absolute frequency of collocation (already adjusted to window size)}

\item{window_size}{total window size around node (left neighbour count + right neighbour count)}
}
\value{
association score
}
\description{
Pointwise mutual information
}
\seealso{
Other association-score-functions: 
\code{\link{defaultAssociationScoreFunctions}()},
\code{\link{ll}()},
\code{\link{logDice}()},
\code{\link{mi2}()},
\code{\link{mi3}()}
}
\concept{association-score-functions}
