% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{bind_cols}
\alias{mutate}
\alias{select}
\alias{group_by}
\alias{summarise}
\alias{n}
\alias{complete}
\alias{expand_grid}
\alias{year}
\alias{as_tibble}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:tidy]{tidy}}}

  \item{dplyr}{\code{\link[dplyr:bind_cols]{bind_cols()}}, \code{\link[dplyr:group_by]{group_by()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:n]{n()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:summarise]{summarise()}}}

  \item{lubridate}{\code{\link[lubridate:year]{year()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}}

  \item{tidyr}{\code{\link[tidyr:complete]{complete()}}, \code{\link[tidyr:expand_grid]{expand_grid()}}}
}}

