% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{show,KorAPConnection-method}
\alias{show,KorAPConnection-method}
\title{Display KorAPConnection object}
\usage{
\S4method{show}{KorAPConnection}(object)
}
\arguments{
\item{object}{KorAPConnection object}
}
\description{
Display KorAPConnection object
}
\keyword{internal}
