% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{clearAccessToken,KorAPConnection-method}
\alias{clearAccessToken,KorAPConnection-method}
\alias{clearAccessToken}
\title{Clear access token from keyring and KorAPConnection object}
\usage{
\S4method{clearAccessToken}{KorAPConnection}(kco)
}
\arguments{
\item{kco}{KorAPConnection object}
}
\value{
KorAPConnection object with access token set to \code{NULL}.
}
\description{
Clear access token from keyring and KorAPConnection object
}
\examples{
\dontrun{
kco <- new("KorAPConnection")
kco <- clearAccessToken(kco)
}
}
\seealso{
\code{\link[=persistAccessToken]{persistAccessToken()}}
}
