% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPCorpusStats.R
\docType{methods}
\name{corpusStats,KorAPConnection-method}
\alias{corpusStats,KorAPConnection-method}
\alias{corpusStats}
\title{Fetch information about a (virtual) corpus}
\usage{
\S4method{corpusStats}{KorAPConnection}(kco, vc = "")
}
\arguments{
\item{kco}{\code{\link{KorAPConnection}} object (obtained e.g. from \code{new("KorAPConnection")}}

\item{vc}{string describing the virtual corpus. An empty string (default) means the whole corpus, as far as it is license-wise accessible.}
}
\value{
\code{KorAPCorpusStats} object with the slots \code{documents}, \code{tokens}, \code{sentences}, \code{paragraphs}
}
\description{
Fetch information about a (virtual) corpus
}
\examples{
corpusStats(new("KorAPConnection"))

kco <- new("KorAPConnection")
corpusStats(kco, "pubDate in 2017 & articleType=/Zeitung.*/")

}
