% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPQuery.R
\name{fetchRest,KorAPQuery-method}
\alias{fetchRest,KorAPQuery-method}
\alias{fetchRest}
\title{Fetches the remaining results of a KorAP query.}
\usage{
\S4method{fetchRest}{KorAPQuery}(kqo, verbose = kqo@korapConnection@verbose, ...)
}
\description{
Fetches the remaining results of a KorAP query.
}
\examples{
\dontrun{

q <- KorAPConnection() \%>\%
  corpusQuery("Ameisenplage") \%>\%
  fetchRest()
q@collectedMatches
}

}
