% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{geom_freq_by_year_ci}
\alias{geom_freq_by_year_ci}
\title{Plot frequency by year graphs with confidence intervals}
\usage{
geom_freq_by_year_ci()
}
\description{
Convenience function for plotting typical frequency by year graphs with confidence intervals using ggplot2.
}
\examples{
library(ggplot2)
kco <- new("KorAPConnection", verbose=TRUE)
expand_grid(condition = c("textDomain = /Wirtschaft.*/", "textDomain != /Wirtschaft.*/"),
            year = (2002:2018)) \%>\%
  cbind(frequencyQuery(kco, "[tt/l=Heuschrecke]",
                            paste0(.$condition," & pubDate in ", .$year)))  \%>\%
  ipm() \%>\%
  ggplot(aes(year, ipm, fill = condition, color = condition, ymin = conf.low, ymax = conf.high)) +
  geom_freq_by_year_ci()

}
