% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggplotly}
\alias{ggplotly}
\title{Convert ggplot2 to plotly with hyperlinks to KorAP queries}
\usage{
ggplotly(p = ggplot2::last_plot(), tooltip = c("x", "y", "url"), ...)
}
\arguments{
\item{p}{a ggplot object.}

\item{tooltip}{a character vector specifying which aesthetic mappings to show
in the tooltip. If you want hyperlinks to KorAP queries you need to include
\code{"url"} here.}

\item{...}{Other arguments passed to \code{plotly::ggplotly}}
}
\description{
This function converts a \code{ggplot2::ggplot()} object to a plotly
objects with hyperlinks from data points to corresponding KorAP queries.
}
\examples{
library(ggplot2)
kco <- new("KorAPConnection", verbose=TRUE)
g <- expand_grid(condition = c("textDomain = /Wirtschaft.*/", "textDomain != /Wirtschaft.*/"),
            year = (2002:2018)) \%>\%
  cbind(frequencyQuery(kco, "[tt/l=Heuschrecke]",
                       paste0(.$condition," & pubDate in ", .$year)))  \%>\%
  ipm() \%>\%
  ggplot(aes(year, ipm, fill = condition, color = condition)) +
##  theme_light(base_size = 20) +
  geom_freq_by_year_ci()
p <- ggplotly(g)
print(p)
## saveWidget(p, paste0(tmpdir(), "heuschrecke.html")


}
