% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-helper.R
\name{hc_add_onclick_korap_search}
\alias{hc_add_onclick_korap_search}
\title{Add KorAP search click events to highchart}
\usage{
hc_add_onclick_korap_search(hc)
}
\arguments{
\item{hc}{highchart}
}
\description{
Adds on-click events to data points of highcarts that were constructed with
\ref{frequencyQuery} or ref \ref{collocationScoreQuery}. Clicks on data points
then launch KorAP web UI queries for the given query term and virtual corpus in
a separate frame.
}
\examples{
\donttest{
library(highcharter)
new("KorAPConnection", verbose = TRUE) \%>\%
  collocationScoreQuery("Team", "agil", vc = paste("pubDate in", c(2014:2018)),
                        lemmatizeNodeQuery = TRUE, lemmatizeCollocateQuery = TRUE) \%>\%
                         pivot_longer(c("O", "E")) \%>\%
  hchart(type="spline", hcaes(label, score, group=name)) \%>\%
  hc_add_onclick_korap_search()
}

}
