% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_freq_by_year_ci.R
\name{hc_freq_by_year_ci}
\alias{hc_freq_by_year_ci}
\title{Plot interactive frequency curves with confidence intervals}
\usage{
hc_freq_by_year_ci(
  df,
  as.alternatives = FALSE,
  ylabel = if (as.alternatives) "\%" else "ipm",
  smooth = FALSE,
  ...
)
}
\arguments{
\item{df}{data frame like the value of a \code{\link[=frequencyQuery]{frequencyQuery()}}}

\item{as.alternatives}{boolean decides whether queries should be treated as mutually exclusive and exhaustive wrt. to some meaningful class (e.g. spelling variants of a certain word form).}

\item{ylabel}{defaults to \verb{\%} if \code{as.alternatives} is \code{TRUE} and to \code{ipm} otherwise.}

\item{smooth}{boolean decides whether the graph is smoothed using the highcharts plot types spline and areasplinerange.}

\item{...}{additional arguments passed to \code{\link[=hc_add_series]{hc_add_series()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Convenience function for plotting typical frequency by year graphs with confidence intervals using highcharter.

\strong{Warning:} This function may be moved to a new package.
}
\seealso{
Other highcharter-helpers: 
\code{\link{hc_add_onclick_korap_search}()}
}
\concept{highcharter-helpers}
