% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\docType{class}
\name{KorAPConnection-class}
\alias{KorAPConnection-class}
\alias{KorAPConnection}
\alias{initialize,KorAPConnection-method}
\alias{show,KorAPConnection-method}
\title{Class KorAPConnection}
\usage{
\S4method{initialize}{KorAPConnection}(.Object,
  KorAPUrl = "https://korap.ids-mannheim.de/", apiVersion = "v1.0",
  apiUrl)

\S4method{show}{KorAPConnection}(object)
}
\arguments{
\item{.Object}{KorAPConnection object}

\item{KorAPUrl}{the URL of the KorAP server instance you want to access.}

\item{apiVersion}{which version of KorAP's API you want to connect to.}

\item{apiUrl}{URL of the KorAP web service.}

\item{object}{KorAPConnection object}
}
\value{
\code{\link{KorAPConnection}} object that can be used e.g. with \code{\link{corpusQuery}}
}
\description{
\code{KorAPConnection} objetcs represent the connection to a KorAP server.
New \code{KorAPConnection} objects can be created by \code{KorAPConnection()}
}
\note{
Currently it is not possible to authenticate the client
}
\examples{
kcon <- new("KorAPConnection")
kq <- corpusQuery(kcon, "Ameisenplage")
kq <- fetchAll(kq, verbose=TRUE)

}
