% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{KorAPConnection}
\alias{KorAPConnection}
\title{\code{KorAPQuery} initiates a connect to some KorAP server.}
\usage{
KorAPConnection(KorAPUrl = defaultKorAPUrl, apiVersion = "v1.0",
  apiUrl = NA)
}
\arguments{
\item{KorAPUrl}{instead of providing the query and vc string parameters, you can also simply copy a KorAP query URL from your browser and use it here (and in \code{KorAPConnection}) to provide all necessary information for the query.}
}
\value{
object that contains all necessary connection information and can be used on \code{\link{KorAPQuery}}
}
\description{
\code{KorAPQuery} initiates a connect to some KorAP server.
}
