% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPConnection.R
\name{KorAPConnection}
\alias{KorAPConnection}
\title{Connect to a KorAP server.}
\usage{
KorAPConnection(KorAPUrl = defaultKorAPUrl, apiVersion = "v1.0",
  apiUrl = NA)
}
\arguments{
\item{KorAPUrl}{instead of providing the query and vc string parameters, you can also simply copy a KorAP query URL from your browser and use it here (and in \code{KorAPConnection}) to provide all necessary information for the query.}
}
\value{
object that contains all connection information and can be used with \code{\link{KorAPQuery}}
}
\description{
Connect to a KorAP server.
}
\note{
Currently it is not possible to authenticate the client
}
