% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KorAPCorpusStats.R
\docType{class}
\name{KorAPCorpusStats-class}
\alias{KorAPCorpusStats-class}
\alias{show,KorAPCorpusStats-method}
\title{KorAPCorpusStats class (internal)}
\usage{
\S4method{show}{KorAPCorpusStats}(object)
}
\arguments{
\item{object}{KorAPCorpusStats object}
}
\description{
Internal class for corpus statistics storage. Users work with \code{corpusStats()} function instead.
}
\keyword{internal}
