% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocationScoreQuery.R
\name{mergeDuplicateCollocates}
\alias{mergeDuplicateCollocates}
\title{Merge duplicate collocate rows and re-calculate association scores and urls}
\usage{
mergeDuplicateCollocates(...)
}
\arguments{
\item{...}{tibbles with collocate rows returned from \code{\link[=collocationAnalysis]{collocationAnalysis()}}}
}
\value{
tibble with unique collocate rows
}
\description{
Merge duplicate collocate rows and re-calculate association scores and urls
}
