% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{percent}
\alias{percent}
\title{Convert corpus frequency table of alternatives to percent}
\usage{
percent(df)
}
\arguments{
\item{df}{table returned from \code{\link{frequencyQuery}}}
}
\value{
original table with converted columns \code{f}, \code{conf.low} and \code{conf.high}
}
\description{
Convenience function for converting frequency tables of alternative variants
(generated with \code{as.alternatives=T}) to percent.
}
\examples{
new("KorAPConnection") \%>\%
    frequencyQuery(c("Tollpatsch", "Tolpatsch"),
    paste0("pubDate in ", 2000:2002)
    as.alternatives = T,
    )
    \%>\% percent()
}
