% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKorAPClient-package.R
\docType{package}
\name{RKorAPClient-package}
\alias{RKorAPClient}
\alias{RKorAPClient-package}
\title{RKorAPClient: 'KorAP' Web Service Client Package}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A client package that makes the 'KorAP' web service API accessible from R. The corpus analysis platform 'KorAP' has been developed as a scientific tool to make potentially large, stratified and multiply annotated corpora, such as the 'German Reference Corpus DeReKo' or the 'Corpus of the Contemporary Romanian Language CoRoLa', accessible for linguists to let them verify hypotheses and to find interesting patterns in real language use. The 'RKorAPClient' package provides access to 'KorAP' and the corpora behind it for user-created R code, as a programmatic alternative to the 'KorAP' web user-interface. You can learn more about 'KorAP' and use it directly on 'DeReKo' at <https://korap.ids-mannheim.de/>.
}
\references{
Kupietz, Marc / Diewald, Nils / Margaretha, Eliza (2020):
RKorAPClient: An R package for accessing the German Reference Corpus DeReKo
via KorAP. In: Calzolari, Nicoletta, Frédéric Béchet, Philippe Blache,
Khalid Choukri, Christopher Cieri,  Thierry Declerck, Sara Goggi,
Hitoshi Isahara, Bente Maegaard, Joseph Mariani, Hélène Mazo,
Asuncion Moreno, Jan Odijk, Stelios Piperidis (eds.):
Proceedings of The 12th Language Resources and Evaluation Conference (LREC 2020)
Marseille: European Language Resources Association (ELRA), 7017-7023.
\url{http://www.lrec-conf.org/proceedings/lrec2020/pdf/2020.lrec-1.867.pdf}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KorAP/RKorAPClient/}
  \item \url{https://korap.ids-mannheim.de/}
  \item \url{https://www.ids-mannheim.de/digspra/kl/projekte/korap}
  \item Report bugs at \url{https://github.com/KorAP/RKorAPClient/issues}
}

}
\author{
\strong{Maintainer}: Marc Kupietz \email{kupietz@ids-mannheim.de}

Other contributors:
\itemize{
  \item Nils Diewald \email{diewald@ids-mannheim.de} [contributor]
  \item Leibniz Institute for the German Language [copyright holder, funder]
}

}
\keyword{internal}
