% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RKorAPClient.R
\docType{package}
\name{RKorAPClient}
\alias{RKorAPClient}
\alias{RKorAPClient-package}
\title{\code{RKorapClient} package}
\description{
R pakcage to access the \href{https://github.com/KorAP/}{KorAP} web service API.
}
\details{
See the README.md on
\href{https://github.com/KorAP/RKorAPClient/}{github}
}
