% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocationAnalysis.R
\name{synsemanticStopwords}
\alias{synsemanticStopwords}
\title{Preliminary synsemantic stopwords function}
\usage{
synsemanticStopwords(...)
}
\arguments{
\item{...}{future arguments for language detection}
}
\value{
Vector of synsemantic stopwords.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Preliminary synsemantic stopwords function to be used in collocation analysis.
}
\details{
Currently only suitable for German See \link[stopwords]{stopwords} for other languages.
}
\seealso{
Other collocation analysis functions: 
\code{\link{association-score-functions}},
\code{\link{collocationAnalysis,KorAPConnection-method}},
\code{\link{collocationScoreQuery,KorAPConnection-method}}
}
\concept{collocation analysis functions}
